frontend: components restructured
