+++ frontend/src/App.vue
@@ -9,8 +9,8 @@
 </template>
 
 <script>
-import HeaderComponent from './components/HeaderComponent.vue'
-import FooterComponent from '@/components/FooterComponent.vue'
+import HeaderComponent from '@/components/header/HeaderComponent.vue'
+import FooterComponent from '@/components/footer/FooterComponent.vue'
 
 export default {
   name: 'App',
