+++ frontend/src/components/books/BookCardComponent.vue
