+++ frontend/src/components/books/BookReadComponent.vue
