+++ frontend/src/components/footer/FooterComponent.vue
