+++ frontend/src/components/general/LoaderComponent.vue
