+++ frontend/src/components/header/NavBarComponent.vue
