+++ frontend/src/components/header/TopBarComponent.vue
